examples/experiments/a2ui-private/original_a2ui_source/a2a_samples/a2ui_restaurant_finder/pyproject.toml
[project]
name = "a2ui-restaurant-finder"
version = "0.1.0"
description = "Sample Google ADK-based Restaurant finder agent that uses a2ui UI and is hosted as an A2A server agent."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "google-adk>=1.8.0",
    "google-genai>=1.27.0",
    "python-dotenv>=1.1.0",
    "litellm",
    "a2ui_ext",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
a2ui_ext = { workspace = true }
